# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fly Delivery",
    "description": "This application offers a seamless and convenient way to satisfy your cravings with just a few clicks. Fly-Delivery connects you with a wide range of restaurants and cuisines, bringing delicious meals straight to your doorstep."
  },
  "servers": [
    {
      "url": "http://localhost:4444/"
    }
  ],
...
  "paths": {
...
    "/orders/{id}": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "orders"
        ],
        "summary": "Add new Order",
        "description": "Add new Order",
        "operationId": "AddOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The article ID",
            "required": true
          }
        ],
        "requestBody": {
          "description": "Add new Order",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numberPieces": {
                    "type": "integer",
                    "example": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Article created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "404": {
            "description": "Bad request!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      },
      "patch": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "orders"
        ],
        "summary": "You can change number of pieces",
        "description": "Change number of pieces",
        "operationId": "changeNbrPieces",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The order that needs to updated",
            "required": true
          }
        ],
        "requestBody": {
          "description": "Order with new number of pieces",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numberPieces": {
                    "type": "integer",
                    "example": 2
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200":{
            "description": "Order updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "numberPieces": {
                      "type": "integer",
                      "example": 2
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "orders"
        ],
        "summary": "Delete order",
        "description": "Delete order",
        "operationId": "deleteOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The order that needs to be deleted",
            "required": true
          }
        ], 
        "responses": {
          "200": {
            "description": "Order is successfully deleted"
          },
          "400": {
            "description": "Bad request!"
          },
          "404": {
            "description": "Order not found!"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "orders"
        ],
        "summary": "Get all orders and orders created by the same user OR received at the same restaurant OR available to the deliveryman",
        "description": "Get all orders and orders created by the same user OR received at the same restaurant OR available to the deliveryman",
        "operationId": "getAllOrders",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of orders per page",
            "required": true,
            "type": "integer"
          },
          {
            "name": "sortBy (fill in by admin)",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "type": "string"
          },
          {
            "name": "createdAtBefore (fill in by admin)",
            "in": "query",
            "description": "Filter orders created before this date (YYYY-MM-DD format)",
            "required": false,
            "type": "string"
          },
          {
            "name": "createdAtAfter (fill in by admin)",
            "in": "query",
            "description": "Filter orders created after this date (YYYY-MM-DD format)",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "exemple": true
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 10
                    },
                    "listOrders": {
                      "type": "integer",
                      "example": 1
                    },
                    "orders": {
                      "type": "array",
                      "example": [
                        {
                        "_id": "66aa509cbe3698c873106862",
                        "articleId": "666b776b78b0808b9a7d8d70",
                        "clientId": "666b76561c55c46db252635f",
                        "restaurantId": "666b769c464843e905f67165",
                        "restaurantStatus": null,
                        "deliverymanStatus": false,
                        "orderStatus": false,
                        "numberPieces": 1,
                        "pricePieces": 9000,
                        "priceOrder": 9000,
                        "createdAt": "2024-07-31T14:56:28.211Z",
                        "updatedAt": "2024-07-31T14:56:28.211Z",
                        "__v": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "No orders yet in this period"
          },
          "404": {
            "description": "Orders not found!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "orders"
        ],
        "summary": "Delete all orders created by the same customer",
        "description": "Delete all orders created by the same customer",
        "operationId": "deleteOrdersCustomer",
        "responses": {
          "200": {
            "description": "All your orders are deleted"
          },
          "404": {
            "description": "You have not yet an order!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      }
    },
    "/orders/restaurant/{id}": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "orders"
        ],
        "summary": "Make the order decision by the restaurant",
        "description": "Make the order decision by the restaurant",
        "operationId": "orderDecisionRestaurant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The order that needs to answer",
            "required": true
          }
        ],
        "requestBody": {
          "description": "Decision of order",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Your answer is sent successfully"
          },
          "404": {
            "description": "Order not found!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      },
      "patch": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "orders"
        ],
        "summary": "Change the order decision by the restaurant",
        "description": "Change the order decision by the restaurant",
        "operationId": "changeOrderDecision",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The order that needs to change the decision",
            "required": true
          }
        ],
        "requestBody": {
          "description": "Change decision of order",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Your answer is sent successfully"
          },
          "400": {
            "description": "Your cannot change your decision!"
          },
          "404": {
            "description": "Order not found!"
          },
          "500": {
            "description": "Internal Server Error!"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Article": {
        "required": [
          "restaurantId",
          "name",
          "price",
          "information"
        ],
        "type": "object",
        "properties": {
          "restaurantId": {
            "type": "string",
            "example": "66063369f5a2397ad3b4fa5d"
          },
          "name": {
            "type": "string",
            "example": "Pizza"
          },
          "picture": {
            "type": "string",
            "example": "picture of this article"
          },
          "price": {
            "type": "integer",
            "example": 20
          },
          "rating": {
            "type": "integer",
            "example": 1
          },
          "reviews": {
            "type": "integer",
            "example": 10
          },
          "discount": {
            "type": "integer",
            "example": 30
          },
          "information": {
            "type": "string",
            "example": "all information about this article"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}